+++ src/reactpy_django/__init__.py
@@ -1,3 +1,7 @@
+import contextlib
+
+import nest_asyncio
+
 from reactpy_django import checks, components, decorators, hooks, types, utils
 from reactpy_django.websocket.paths import REACTPY_WEBSOCKET_PATH
 
@@ -11,3 +15,9 @@ __all__ = [
     "utils",
     "checks",
 ]
+# Built-in asyncio event loops can create `assert f is self._write_fut` exceptions
+# while we are using our backhaul thread with Uvicorn, so we use this patch to fix this.
+# This also resolves jittery rendering behaviors within Daphne. Can be demonstrated
+# using our "Renders Per Second" test page.
+with contextlib.suppress(ValueError):
+    nest_asyncio.apply()