+++ CHANGELOG.md
@@ -36,6 +36,19 @@ Using the following categories, list your changes in this order:
 
 -   Nothing (yet)!
 
+## [3.3.2] - 2023-08-13
+
+### Added
+
+-   ReactPy Websocket will now decode messages via `orjson` resulting in an ~6% overall performance improvement.
+-   Built-in `asyncio` event loops are now patched via `nest_asyncio`, resulting in an ~10% overall performance improvement. This has no performance impact if you are running your webserver with `uvloop`.
+
+### Fixed
+
+-   Fix bug where `REACTPY_WEBSOCKET_URL` always generates a warning if unset.
+-   Fixed bug on Windows where `assert f is self._write_fut` would be raised by `uvicorn` when `REACTPY_BACKHAUL_THREAD = True`.
+-   Fixed bug on Windows where rendering behavior would be jittery with `daphne` when `REACTPY_BACKHAUL_THREAD = True`.
+
 ## [3.3.1] - 2023-08-08
 
 ### Added